Send values based on the scanning ability of a pass

+++ zx_scanner/app/send_event.h
@@ -14,7 +14,7 @@ extern "C" {
 typedef struct {
     AXEventHandler* event_handler;
     guint event_id;
-    gdouble value;
+    gint value;
 } AppData;
 
 // Function declarations
